Working through feedback, onto patch

+++ db/migrations/20200504124415_create_articles_table.js
@@ -3,7 +3,7 @@ exports.up = function (knex) {
     articlesTable.increments("article_id").primary();
     articlesTable.string("title").notNullable();
     articlesTable.text("body").notNullable();
-    articlesTable.bigInteger("votes").defaultTo(0);
+    articlesTable.integer("votes").defaultTo(0);
     articlesTable.text("topic").references("topics.slug").notNullable();
     articlesTable.string("author").references("users.username").notNullable();
     articlesTable.timestamp("created_at").defaultTo(knex.fn.now());
