+++ db/migrations/20200504133702_create_comments_table.js
@@ -7,7 +7,7 @@ exports.up = function (knex) {
       .references("articles.article_id")
       .notNullable()
       .onDelete("CASCADE");
-    commentsTable.bigInteger("votes").defaultTo(0);
+    commentsTable.integer("votes").defaultTo(0);
     commentsTable.timestamp("created_at").defaultTo(knex.fn.now());
     commentsTable.text("body");
   });
